+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 * Support an unlimited number of cached filters, instead of 128 (#245)
 * `WorldStats` contains the number of cached filters (#247)
 * Archetypes with entity relations are removed on `World.Reset` (#247)
+* Capacity increment can be configured separately for relation archetypes (#257)
 
 ### Performance